+++ src/features/Home/components/RandomCharacter/index.js
@@ -12,17 +12,25 @@ import Tooltip from 'common/components/Tooltip';
 
 import styles from './styles.less';
 
-/* eslint max-len:0 */
+
 export default class RandomCharacter extends Component {
+  props: {
+    type: 'ofTheDay' | 'random',
+  };
+
   state = {
     name: '',
   };
 
   componentDidMount () {
-    axios.get(`${config.api.endpoint}random/characters/1`, {
+    const resource = this.props.type === 'ofTheDay'
+      ? 'of-the-day/characters'
+      : 'random/characters/1';
+
+    axios.get(`${config.api.endpoint}${resource}`, {
       ignoreAuth: true,
     })
-      .then(({ data }) => this.setState({ name: data[0] }));
+    .then(({ data }) => this.setState({ name: data[0] }));
   }
 
   render () {