+++ src/features/Home/components/RandomGuilds/index.js
@@ -2,13 +2,14 @@
 
 import axios from 'axios';
 import { Component } from 'react';
-import ContentCard from 'common/components/ContentCard';
-import { Link } from 'react-router';
+import GuildContentCard from 'common/components/ContentCard/Guild';
 import get from 'lodash/get';
 
 import config from 'config';
 
 import styles from './styles.less';
+
+const STUB_GUILDS = [undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined];
 
 export default class RandomGuilds extends Component {
   state = {
@@ -16,22 +17,20 @@ export default class RandomGuilds extends Component {
   };
 
   componentDidMount () {
-    axios.get(`${config.api.endpoint}random/guilds/8`, {
+    axios.get(`${config.api.endpoint}of-the-day/guilds`, {
       ignoreAuth: true,
     })
-      .then(({ data }) => this.setState({ guilds: data }));
+    .then(({ data }) => this.setState({ guilds: data }));
   }
 
   render () {
-    const guilds = get(this.state, 'guilds', [undefined, undefined, undefined, undefined, undefined, undefined, undefined, undefined]);
+    const guilds = get(this.state, 'guilds', STUB_GUILDS);
 
     return (
       <ul className={styles.root}>
         {guilds.map((guild, index) => (
           <li key={guild ? guild.name : index}>
-            <Link to={`g/${guild && guild.name}`}>
-              <ContentCard type="guilds" content={guild} />
-            </Link>
+            <GuildContentCard type="guilds" content={guild} />
           </li>
         ))}
       </ul>