+++ src/features/Leaderboards/actions.js
@@ -2,6 +2,7 @@
 
 import { get } from 'axios';
 import config from 'config';
+import { paginatedThunk } from 'lib/redux';
 
 export const FETCH_PVP_LEADERBOARD = 'FETCH_PVP_LEADERBOARD';
 
@@ -13,8 +14,15 @@ const fetchPvpLeaderboardSuccess = (data, region) => ({
   type: FETCH_PVP_LEADERBOARD,
 });
 
-export function fetchPvpLeaderboard (region: 'na' | 'eu' | 'gw2a') {
-  return (dispatch: Dispatch) => get(`${config.api.endpoint}leaderboards/pvp/${region}`, {
-    ignoreAuth: true,
-  }).then(({ data }) => dispatch(fetchPvpLeaderboardSuccess(data, region)));
+export function fetchPvpLeaderboard (region: 'na' | 'eu' | 'gw2a', limit: number, offset: number) {
+  return paginatedThunk((dispatch: Dispatch) => {
+    return get(`${config.api.endpoint}leaderboards/pvp/${region}`, {
+      ignoreAuth: true,
+      params: {
+        limit,
+        offset,
+      },
+    })
+    .then(({ data }) => dispatch(fetchPvpLeaderboardSuccess(data, region)));
+  }, `leaderboards.pvp[${region}]`, limit, offset);
 }