+++ src/features/User/index.js
@@ -1,6 +1,9 @@
 // @flow
 
-import { Component, PropTypes } from 'react';
+import type { User as UserType, PvpSeasons, Worlds, Maps } from 'flowTypes';
+
+import { Component } from 'react';
+import PropTypes from 'prop-types';
 import { createSelector } from 'reselect';
 import { connect } from 'react-redux';
 import get from 'lodash/get';
@@ -8,19 +11,22 @@ import isObject from 'lodash/isObject';
 import filter from 'lodash/filter';
 import T from 'i18n-react';
 import { Link } from 'react-router';
+import times from 'lodash/times';
 import startCase from 'lodash/startCase';
 
+import PaginatorGrid from 'common/layouts/PaginatorGrid';
 import TooltipTrigger from 'common/components/TooltipTrigger';
 import Button from 'common/components/Button';
 import Icon from 'common/components/Icon';
 import Content from 'common/layouts/Content';
-import ContentCardList from 'common/components/ContentCardList';
+import GuildContentCard from 'common/components/ContentCard/Guild';
 
 import styles from './styles.less';
 import PvpGame from './components/PvpGame';
 import Overview from './components/Overview';
+import Characters from './components/Characters';
 
-import type { User as UserType, PvpSeasons, Worlds, Maps } from 'flowTypes';
+const STUB_GUILDS = times(4, () => undefined);
 
 import {
   fetchUser,
@@ -60,6 +66,7 @@ function getActiveStanding ({ pvpStandings = [] } = {}, pvpSeasons) {
 }
 
 const addHash = (str) => (str ? `#${str}` : '-');
+const makeKey = (content, index) => (content ? content.name : index);
 
 @connect(selector, {
   dispatchFetchUser: fetchUser,
@@ -171,25 +178,20 @@ export default class User extends Component {
         }, {
           to: `/${alias}/characters`,
           name: 'Characters',
-          content: (
-            <ContentCardList
-              noBorder
-              type="grid"
-              alias={alias}
-              items={user && user.characters}
-            />
-          ),
+          content: <Characters alias={alias} />,
         }, {
           to: `/${alias}/guilds`,
           name: T.translate('guilds.name'),
           content: (
-            <ContentCardList
-              noBorder
-              type="grid"
-              alias={alias}
-              resource="guilds"
-              items={guilds}
-            />
+            <PaginatorGrid
+              key="guilds"
+              rows={guilds || STUB_GUILDS}
+              limit={0}
+              count={0}
+              action={() => Promise.resolve()}
+            >
+              {(content, index) => <GuildContentCard key={makeKey(content, index)} content={content} />}
+            </PaginatorGrid>
           ),
         }, {
           to: `/${alias}/matches`,